// src/app/accommodation/models/accommodation.ts
export class Accommodation {
  _id: String;
  title: String;
  description: String;
  town: String;
  thumbnail: String;
  avatar: String;
  type: String;
  numberOfGuests: number;
  price: number;
  lat: number;
  lng: number;
  numberOfBeds: number;
  numberOfBathrooms: number;
  zipcode: String;
  canton: String;
  country: String;
  category: String;
  setting: {};
  rules: {};
  rulesText: String;
  headerImage: String;
  images: number;
  username: String;
  distance: number;
  sel: boolean;
  ownerid: string;
}